helpers: do not return string view of a stack var
Returning a view of local std::string("") is not valid. Return an empty
view instead.

It looks like this function returned std::string initially.

Also take a ref to a passed line, so that it is not copied to stack.

Also no need to create a new view from line, use substr() of the
existing one.